requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
sendgrid==6.11.0
pydantic-settings==2.7.0
python-dotenv==1.0.1
slowapi==0.1.9
boto3==1.35.99
pdfplumber==0.11.4
pypdf==5.1.0
